refactor the data_clinical_sample.txt Facets Suite update code into a new cBioPortal_utils.py module so that it can be shared between the update portal script and the generate portal script
