+++ bin/update_cBioPortal_data.py
@@ -91,21 +91,21 @@ import argparse
 
 # relative imports, from CLI and from parent project
 if __name__ != "__main__":
-    from .generate_cbioPortal_files import create_file_lines
+    from .cBioPortal_utils import create_file_lines
+    from .cBioPortal_utils import update_sample_data
+    from .cBioPortal_utils import sample_data_remove_cols
+    from .cBioPortal_utils import facets_data_keep_cols_map
+    from .cBioPortal_utils import parse_facets_data
 
 if __name__ == "__main__":
-    from generate_cbioPortal_files import create_file_lines
+    from cBioPortal_utils import create_file_lines
+    from cBioPortal_utils import update_sample_data
+    from cBioPortal_utils import sample_data_remove_cols
+    from cBioPortal_utils import facets_data_keep_cols_map
+    from cBioPortal_utils import parse_facets_data
 
-# keep only these columns, and rename them to the listed values
-facets_data_keep_cols_map = { # old:new
-    "purity": "ASCN_PURITY",
-    "ploidy": "ASCN_PLOIDY",
-    "facets_version": "ASCN_VERSION",
-    "genome_doubled": "ASCN_WGD"
-}
-
-# remove these columns from data_clinical_sample.txt data while updating it with Facets Suite data
-sample_data_remove_cols = ["purity", "ploidy", "facets_version"]
+# # remove these columns from data_clinical_sample.txt data while updating it with Facets Suite data
+# sample_data_remove_cols = ["purity", "ploidy", "facets_version"]
 mutation_data_keep_cols_map = {
     "tcn": "ASCN.TOTAL_COPY_NUMBER",
     "lcn": "ASCN.MINOR_COPY_NUMBER",
@@ -163,61 +163,6 @@ facets_call_states = {
 ('FALSE', 5,3):2,
 ('FALSE', 6,3):2,
 }
-
-def parse_facets_data(rows):
-    """
-    Need to dig through the data output by Facets Suite .txt files in order to return a dict of per sample info
-    that has only the desired info from "hisens" rows
-    """
-    data = {}
-    for row in rows:
-        # only keep 'hisens' rows
-        if row['run_type'] != "hisens":
-            continue
-        else:
-            # keep only desired columns
-            d = { key:row[key] for key in facets_data_keep_cols_map.keys() }
-            # add the renamed columns
-            for old_key, new_key in facets_data_keep_cols_map.items():
-                d[new_key] = d[old_key]
-            # change the "genome_doubled" value
-            if d["genome_doubled"] == "TRUE":
-                d["ASCN_WGD"] = "WGD"
-            elif d["genome_doubled"] == "FALSE":
-                d["ASCN_WGD"] = "no WGD"
-
-            # NOTE:
-            # convention earlier in the pipeline was that Facets Suite "sample_id" = TumorID.NormalID
-            # we need to use only the Tumor ID here because that is what matches up with the values in
-            # data_clinical_sample.txt
-            # so, split on '.' here an keep the first item and assume its the Tumor ID
-            # TODO: get Facets Suite run-facets-wrapper.R script to embed the tumor and normal ID values into the file
-            # so we dont have to do this
-            sample_id = row['sample'].split('.')[0]
-            data[sample_id] = d
-    return(data)
-
-
-
-def update_sample_data(sample_data, facets_data):
-    """
-    Add entries to the sample data based on values from the facets data
-    """
-    d = { k:v for k, v in sample_data.items() }
-    sample_id = d['SAMPLE_ID']
-    # add matching facets data
-    if sample_id in facets_data:
-        d = { **d, **facets_data[sample_id] }
-    else:
-        # add NA for the facets columns instead
-        for k,v in facets_data_keep_cols_map.items():
-            d[k] = "NA"
-            d[v] = "NA"
-    # get rid of undesired columns
-    for key in sample_data_remove_cols:
-        if key in d:
-            d.pop(key)
-    return(d)
 
 def update_mutation_data(mut_data, facets_data = None, sample_id = None):
     """
