+++ tests/test_update_cBioPortal_data.py
@@ -18,306 +18,10 @@ if __name__ == "__main__":
 THIS_DIR = os.path.dirname(os.path.realpath(__file__))
 PARENT_DIR = os.path.dirname(THIS_DIR)
 sys.path.insert(0, PARENT_DIR)
-from bin.update_cBioPortal_data import parse_facets_data
-from bin.update_cBioPortal_data import update_sample_data
 from bin.update_cBioPortal_data import update_mutation_data
 sys.path.pop(0)
 
 class TestGenerateCBioFiles(unittest.TestCase):
-    def test_parse_facets_data1(self):
-        """
-        test that a list of dicts output from Facets Suite can be loaded correctly
-        should return a dict with only "hisens" data
-        This simulates how data should be loaded from the Facets Suite .txt output files
-        """
-        facets_data = [
-            {
-                'hypoploid': 'FALSE',
-                'run_type': 'purity',
-                'fraction_loh': '0.21',
-                'facets_version': '0.5.14',
-                'hrd_loh': '1',
-                'snp_nbhd': '250',
-                'genome_doubled': 'TRUE',
-                'ndepth': '35',
-                'dipLogR': '-1.1',
-                'sample': 'Tumor1.Normal1',
-                'lst': '1',
-                'seed': '1000',
-                'flags': 'mafR not sufficiently small',
-                'purity': '0.35',
-                'input_file': 'Tumor1.Normal1.snp_pileup.gz',
-                'ntai': '2',
-                'fraction_cna': '1',
-                'min_nhet': '25',
-                'ploidy': '8.1',
-                'cval': '100',
-                'genome': 'hg19'
-            },
-            {
-                'hypoploid': 'FALSE',
-                'run_type': 'hisens',
-                'fraction_loh': '0.22',
-                'facets_version': '0.5.14',
-                'hrd_loh': '3',
-                'snp_nbhd': '250',
-                'genome_doubled': 'TRUE',
-                'ndepth': '35',
-                'dipLogR': '-1.1',
-                'sample': 'Tumor1.Normal1',
-                'lst': '5',
-                'seed': '1000',
-                'flags': '',
-                'purity': '0.36',
-                'input_file': 'Tumor1.Normal1.snp_pileup.gz',
-                'ntai': '6',
-                'fraction_cna': '0.92',
-                'min_nhet': '25',
-                'ploidy': '7.9',
-                'cval': '50',
-                'genome': 'hg19'
-            },
-            {
-                'hypoploid': 'FALSE',
-                'run_type': 'purity',
-                'fraction_loh': '0.21',
-                'facets_version': '0.5.14',
-                'hrd_loh': '1',
-                'snp_nbhd': '250',
-                'genome_doubled': 'FALSE',
-                'ndepth': '35',
-                'dipLogR': '-1.1',
-                'sample': 'Tumor2.Normal2',
-                'lst': '1',
-                'seed': '1000',
-                'flags': 'mafR not sufficiently small',
-                'purity': '0.35',
-                'input_file': 'Tumor2.Normal2.snp_pileup.gz',
-                'ntai': '2',
-                'fraction_cna': '1',
-                'min_nhet': '25',
-                'ploidy': '8.1',
-                'cval': '100',
-                'genome': 'hg19'
-            },
-            {
-                'hypoploid': 'FALSE',
-                'run_type': 'hisens',
-                'fraction_loh': '0.22',
-                'facets_version': '0.5.14',
-                'hrd_loh': '3',
-                'snp_nbhd': '250',
-                'genome_doubled': 'FALSE',
-                'ndepth': '35',
-                'dipLogR': '-1.1',
-                'sample': 'Tumor2.Normal2',
-                'lst': '5',
-                'seed': '1000',
-                'flags': '',
-                'purity': '0.36',
-                'input_file': 'Tumor2.Normal2.snp_pileup.gz',
-                'ntai': '6',
-                'fraction_cna': '0.92',
-                'min_nhet': '25',
-                'ploidy': '7.9',
-                'cval': '50',
-                'genome': 'hg19'
-            }
-            ]
-        parsed_data = parse_facets_data(facets_data)
-        expected_data = {
-        'Tumor1': {
-            "purity": "0.36",
-            "ploidy" : '7.9',
-            "facets_version": '0.5.14',
-            "genome_doubled": 'TRUE',
-            "ASCN_PURITY": "0.36",
-            "ASCN_PLOIDY": '7.9',
-            "ASCN_VERSION": '0.5.14',
-            "ASCN_WGD": "WGD"
-            },
-        'Tumor2': {
-            "purity": "0.36",
-            "ploidy" : '7.9',
-            "facets_version": '0.5.14',
-            "genome_doubled": 'FALSE',
-            "ASCN_PURITY": "0.36",
-            "ASCN_PLOIDY": '7.9',
-            "ASCN_VERSION": '0.5.14',
-            "ASCN_WGD": "no WGD"
-            }
-        }
-        self.maxDiff = None
-        self.assertDictEqual(parsed_data, expected_data)
-
-
-    def test_update_sample_data(self):
-        """
-        Tests for updating the data from data_clinical_sample.txt file with corresponding data loaded from the Facets Suite output .txt file
-        """
-        facets_data = {
-        'Tumor1': {
-            "purity": "0.36",
-            "ploidy" : '7.9',
-            "facets_version": '0.5.14',
-            "genome_doubled": 'TRUE',
-            "ASCN_PURITY": "0.36",
-            "ASCN_PLOIDY": '7.9',
-            "ASCN_VERSION": '0.5.14',
-            "ASCN_WGD": "WGD"
-            },
-        'Tumor2': {
-            "purity": "0.36",
-            "ploidy" : '7.9',
-            "facets_version": '0.5.14',
-            "genome_doubled": 'FALSE',
-            "ASCN_PURITY": "0.36",
-            "ASCN_PLOIDY": '7.9',
-            "ASCN_VERSION": '0.5.14',
-            "ASCN_WGD": "no WGD"
-            }
-        }
-        sample_data =  {
-                'TISSUE_SITE': '',
-                'PROJECT_ID': '08390',
-                'SAMPLE_COVERAGE': '108',
-                'ONCOTREE_CODE': 'MEL',
-                'IGO_ID': '08390_G_95',
-                'PIPELINE': 'roslin',
-                'SAMPLE_TYPE': 'Primary',
-                'PATIENT_ID': 'Patient1',
-                'SPECIMEN_PRESERVATION_TYPE': 'FFPE',
-                'COLLAB_ID': 'Tumor-1',
-                'GENE_PANEL': 'IMPACT468+08390_Hg19',
-                'SAMPLE_ID': 'Tumor1',
-                'REQUEST_ID': '08390_G',
-                'PROJECT_PI': 'smithi',
-                'PIPELINE_VERSION': '2.5.7',
-                'REQUEST_PI': 'smithi',
-                'SAMPLE_CLASS': 'Biopsy'
-            }
-        updated_sample_data = update_sample_data(sample_data, facets_data)
-        expected_sample_data = {
-            'TISSUE_SITE': '',
-            'PROJECT_ID': '08390',
-            'SAMPLE_COVERAGE': '108',
-            'ONCOTREE_CODE': 'MEL',
-            'IGO_ID': '08390_G_95',
-            'PIPELINE': 'roslin',
-            'SAMPLE_TYPE': 'Primary',
-            'PATIENT_ID': 'Patient1',
-            'SPECIMEN_PRESERVATION_TYPE': 'FFPE',
-            'COLLAB_ID': 'Tumor-1',
-            'GENE_PANEL': 'IMPACT468+08390_Hg19',
-            'SAMPLE_ID': 'Tumor1',
-            'REQUEST_ID': '08390_G',
-            'PROJECT_PI': 'smithi',
-            'PIPELINE_VERSION': '2.5.7',
-            'REQUEST_PI': 'smithi',
-            'SAMPLE_CLASS': 'Biopsy',
-            "genome_doubled": 'TRUE',
-            "ASCN_PURITY": "0.36",
-            "ASCN_PLOIDY": '7.9',
-            "ASCN_VERSION": '0.5.14',
-            "ASCN_WGD": "WGD"
-        }
-        self.maxDiff = None
-        self.assertDictEqual(updated_sample_data, expected_sample_data)
-
-        sample_data =  {
-            'TISSUE_SITE': '',
-            'PROJECT_ID': '08390',
-            'SAMPLE_COVERAGE': '502',
-            'ONCOTREE_CODE': 'MEL',
-            'IGO_ID': '08390_G_93',
-            'PIPELINE': 'roslin',
-            'SAMPLE_TYPE': 'Primary',
-            'PATIENT_ID': 'Patient2',
-            'SPECIMEN_PRESERVATION_TYPE': 'FFPE',
-            'COLLAB_ID': 'Tumor-2',
-            'GENE_PANEL': 'IMPACT468+08390_Hg19',
-            'SAMPLE_ID': 'Tumor2',
-            'REQUEST_ID': '08390_G',
-            'PROJECT_PI': 'smithi',
-            'PIPELINE_VERSION': '2.5.7',
-            'REQUEST_PI': 'smithi',
-            'SAMPLE_CLASS': 'Biopsy'
-            }
-        updated_sample_data = update_sample_data(sample_data, facets_data)
-        expected_sample_data = {
-            'TISSUE_SITE': '',
-            'PROJECT_ID': '08390',
-            'SAMPLE_COVERAGE': '502',
-            'ONCOTREE_CODE': 'MEL',
-            'IGO_ID': '08390_G_93',
-            'PIPELINE': 'roslin',
-            'SAMPLE_TYPE': 'Primary',
-            'PATIENT_ID': 'Patient2',
-            'SPECIMEN_PRESERVATION_TYPE': 'FFPE',
-            'COLLAB_ID': 'Tumor-2',
-            'GENE_PANEL': 'IMPACT468+08390_Hg19',
-            'SAMPLE_ID': 'Tumor2',
-            'REQUEST_ID': '08390_G',
-            'PROJECT_PI': 'smithi',
-            'PIPELINE_VERSION': '2.5.7',
-            'REQUEST_PI': 'smithi',
-            'SAMPLE_CLASS': 'Biopsy',
-            "genome_doubled": 'FALSE',
-            "ASCN_PURITY": "0.36",
-            "ASCN_PLOIDY": '7.9',
-            "ASCN_VERSION": '0.5.14',
-            "ASCN_WGD": "no WGD"
-        }
-        self.assertDictEqual(updated_sample_data, expected_sample_data)
-
-        # This one does not have a matching Facets data entry; should give NA values instead
-        sample_data = {
-            'TISSUE_SITE': '',
-            'PROJECT_ID': '08390',
-            'SAMPLE_COVERAGE': '500',
-            'ONCOTREE_CODE': 'MEL',
-            'IGO_ID': '08390_G_91',
-            'PIPELINE': 'roslin',
-            'SAMPLE_TYPE': 'Primary',
-            'PATIENT_ID': 'Patient3',
-            'SPECIMEN_PRESERVATION_TYPE': 'FFPE',
-            'COLLAB_ID': 'Tumor-3',
-            'GENE_PANEL': 'IMPACT468+08390_Hg19',
-            'SAMPLE_ID': 'Tumor3',
-            'REQUEST_ID': '08390_G',
-            'PROJECT_PI': 'smithi',
-            'PIPELINE_VERSION': '2.5.7',
-            'REQUEST_PI': 'smithi',
-            'SAMPLE_CLASS': 'Biopsy'
-            }
-        updated_sample_data = update_sample_data(sample_data, facets_data)
-        expected_sample_data = {
-            'TISSUE_SITE': '',
-            'PROJECT_ID': '08390',
-            'SAMPLE_COVERAGE': '500',
-            'ONCOTREE_CODE': 'MEL',
-            'IGO_ID': '08390_G_91',
-            'PIPELINE': 'roslin',
-            'SAMPLE_TYPE': 'Primary',
-            'PATIENT_ID': 'Patient3',
-            'SPECIMEN_PRESERVATION_TYPE': 'FFPE',
-            'COLLAB_ID': 'Tumor-3',
-            'GENE_PANEL': 'IMPACT468+08390_Hg19',
-            'SAMPLE_ID': 'Tumor3',
-            'REQUEST_ID': '08390_G',
-            'PROJECT_PI': 'smithi',
-            'PIPELINE_VERSION': '2.5.7',
-            'REQUEST_PI': 'smithi',
-            'SAMPLE_CLASS': 'Biopsy',
-            "genome_doubled": 'NA',
-            "ASCN_PURITY": "NA",
-            "ASCN_PLOIDY": 'NA',
-            "ASCN_VERSION": 'NA',
-            "ASCN_WGD": "NA"
-        }
-        self.assertDictEqual(updated_sample_data, expected_sample_data)
-
-
     def test_update_mutation_data(self):
         """
         Test that new rows are added correctly to the mutation data file
